+++ tests/test_install.py
@@ -13,21 +13,11 @@
 # So, some in PLATLIB, and one in each of DATA, HEADERS and SCRIPTS.
 
 import os
-import shutil
-from distutils.command.install import SCHEME_KEYS
-from tempfile import mkdtemp
 
-import wheel.pep425tags
-import wheel.tool
-from wheel.install import WheelFile, get_install_paths
+from wheel.install import WheelFile
 
 THISDIR = os.path.dirname(__file__)
 TESTWHEEL = os.path.join(THISDIR, 'test-1.0-py2.py3-none-win32.whl')
-
-
-def test_path():
-    d = get_install_paths('wheel')
-    assert len(d) == len(SCHEME_KEYS)
 
 
 def test_compatibility_tags():
@@ -73,39 +63,6 @@ def test_pick_best():
 
         # assert_equal(
         #     list(map(get_tags, pick_best(cand_wheels, supp, top=False))), supp)
-
-
-def test_install():
-    def check(*path):
-        return os.path.exists(os.path.join(*path))
-
-    def get_supported():
-        return list(wheel.pep425tags.get_supported()) + [('py3', 'none', 'win32')]
-
-    tempdir = mkdtemp()
-    whl = WheelFile(TESTWHEEL, context=get_supported)
-    assert whl.supports_current_python(get_supported)
-    try:
-        locs = {}
-        for key in ('purelib', 'platlib', 'scripts', 'headers', 'data'):
-            locs[key] = os.path.join(tempdir, key)
-            os.mkdir(locs[key])
-        whl.install(overrides=locs)
-        assert len(os.listdir(locs['purelib'])) == 0
-        assert check(locs['platlib'], 'hello.pyd')
-        assert check(locs['platlib'], 'hello', 'hello.py')
-        assert check(locs['platlib'], 'hello', '__init__.py')
-        assert check(locs['data'], 'hello.dat')
-        assert check(locs['headers'], 'hello.dat')
-        assert check(locs['scripts'], 'hello.sh')
-        assert check(locs['platlib'], 'test-1.0.dist-info', 'RECORD')
-    finally:
-        shutil.rmtree(tempdir)
-
-
-def test_install_tool():
-    """Slightly improve coverage of wheel.install"""
-    wheel.tool.install([TESTWHEEL], force=True, dry_run=True)
 
 
 def test_wheelfile_re():